refactor bridge generate wrapped.json
