+++ packages/webpack-plugin/src/templates/components/components.wxml/flatten.tsx
@@ -1,4 +1,4 @@
-import { t } from '../../helper';
+import { t } from '../../helpers/t';
 
 // FIXME: should remove `text` node generated by `components` later
 // FIXME: how to handle more depth of `text` ?
