+++ packages/webpack-plugin/src/templates/components/wrapped.json.ts
@@ -1,12 +1,18 @@
 import { ComponentDesc } from '../../constants/components';
 import { nativeComponentJson } from '../commons/nativeComponentJson';
 
-export const subtreeJson = ({
+export const wrappedJson = ({
   relativePathToBridge,
+  component,
   components,
 }: {
   relativePathToBridge: string;
+  component: ComponentDesc;
   components: ComponentDesc[];
 }) => {
-  return nativeComponentJson({ relativePathToBridge, components, isLeaf: false });
+  return nativeComponentJson({
+    relativePathToBridge,
+    components,
+    isLeaf: component.isLeaf ?? false,
+  });
 };
