--- packages/webpack-plugin/src/templates/context/index.ts
@@ -1,8 +1,0 @@
-import { GojiTarget } from '@goji/core';
-import { createContext } from '../helper';
-
-export interface CommonContextType {
-  target: GojiTarget;
-}
-
-export const CommonContext = createContext<CommonContextType>();
