+++ packages/webpack-plugin/src/templates/helpers/__tests__/t.test.ts
@@ -1,4 +1,4 @@
-import { t, createContext, withContext } from '../helper';
+import { t } from '../t';
 
 describe('template helper `t`', () => {
   test('should trim', () => {
@@ -38,38 +38,5 @@ describe('template helper `t`', () => {
     `).toBe('1\n2\n3\n4');
     // FIXME: refactor `inlineArrayTransformer`
     // expect(t`1\n${[2, 3, 4]}`).toBe('1\n2\n3\n4');
-  });
-});
-
-describe('template context', () => {
-  const testContext = createContext<{ data: number }>();
-  const render = () => t`
-    "data": "${() => testContext.read().data}"
-  `;
-
-  test('should throw if not initialed', () => {
-    expect(() => render()).toThrow();
-  });
-
-  test('should works', () => {
-    expect(withContext(testContext, { data: 123 }, render)).toBe('"data": "123"');
-
-    const nestedRender = () => t`
-      {
-        "json": {
-          ${render}
-        }
-      }`;
-    expect(withContext(testContext, { data: 123 }, nestedRender)).toBe(
-      JSON.stringify(
-        {
-          json: {
-            data: '123',
-          },
-        },
-        null,
-        2,
-      ),
-    );
   });
 });
