+++ packages/webpack-plugin/src/templates/helpers/t.ts
@@ -42,29 +42,3 @@ export const t = createTag(
   // always trim end of result
   trimResultTransformer('end'),
 );
-
-export class Context<T> {
-  private value: T | undefined;
-
-  public read(): T {
-    if (this.value === undefined) {
-      throw new Error('Context must be initialed before using');
-    }
-    return this.value;
-  }
-
-  public write(newValue: T): void {
-    this.value = newValue;
-  }
-}
-
-export function createContext<T>() {
-  return new Context<T>();
-}
-
-export function withContext<T, R>(context: Context<T>, value: T, callback: () => R): R {
-  context.write(value);
-  const result = callback();
-  context.write(value);
-  return result;
-}
