+++ packages/webpack-plugin/src/templates/index.ts
@@ -1,5 +1,4 @@
-import { withContext } from './helper';
-import { CommonContext, CommonContextType } from './context';
+import { withContext, CommonContext, CommonContextType } from './helpers/context';
 
 export const renderTemplate = (contextValue: CommonContextType, callback: () => string) =>
   withContext(CommonContext, contextValue, callback);
