--- packages/webpack-plugin/templates/components/input.json.ejs
@@ -1,4 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {}
-}
