--- packages/webpack-plugin/templates/components/map.json.ejs
@@ -1,4 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {}
-}
