--- packages/webpack-plugin/templates/components/scroll-view.json.ejs
@@ -1,12 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {
-    <% for (const component of components) {%>
-      <% if (component.isWrapped) {%>
-        "goji-<%= component.name %>": "<%= relativePathToBridge %>/<%= component.name %>",
-      <% } %>
-    <% }%>
-
-    "goji-subtree": "<%= relativePathToBridge %>/../subtree"
-  }
-}
