--- packages/webpack-plugin/templates/components/textarea.json.ejs
@@ -1,5 +1,0 @@
-{
-  "component": true,
-  "usingComponents": {}
-}
-
